Add support for MultiLine Text Boxes in GTK for SEG_TEXT.
LF = CTRL+Enter support for direct text entry in the Text Cmd.

When in Describe, tabs and linefeeds are escaped as \t and \n.
Numerous fixes included to ensure the select area and describe boxes are correctly draw and used for multi-line text.
Support for saving and reading escaped text in the SEG_TEXT records.