Make Debug menu work.
This menu in Options->Debug only appears if the env variable XTRKCADEXTRA is set (“Loosen” command also appears in Modify).  Formerly Debug would abend if invoked….!

The Debug window lists any LOGging entries (set with “-d loggingname=level” parms). For example “-d trainMove=5 -d traverseCornu=2” sets two Loglines one at level 5 and the other 2.

The value of the level can be adjusted in the Debug window and then the button “OK” sets it. Given that a level value of 0 means no logging for that logging variable, this menu allows log/tracing to be adjusted on the fly after startup.

Any log entries created before the first invocation of the window will be included, so a coder could add a LogSet(“traverseBezier”,0) line into the InitTrkBezier() code while testing or use a -d traverseBezier=0 and then use Debug to set level to 1 and start logging.