Add full Module support including documentation for modular layouts.
“Module” is a flag setting on a Layer - when it is on it ensures that a Layer is selected (and moved or rotated) always as a group.

Modules can be built by hand, or imported using “Import as Module” from an *.xti file (an Export file). That finds an empty layer, names it for the file, adds all the tracks and sets the Module flag.

The expectation is that a module template might exist as an *.xti file which has the edges and connecting tracks defined. A user would add the contents and export as a different *.xti file in a library of modules. Finally the desired modules would be selected, Imported As Modules, assembled in a new master layout, and operating sessions can be simulated to work out any issues.