Fix #351 Crash loading an existing file
Crash because trk ptr is null in CheckCoupling.
Cause seems to be that a car is overhanging the end of a track and another car
is nearby but not coupled

Fix is not to try coupling if the next car is null.

Somehow the CAR_STATE_IGNORED bit is set which makes the car unselectable.
The car has to be removed from the layout and re-placed.