BUG: crash on zooming in on a large layout with a background bitmap on Windows I see 2 things.
When I zoom in to 1:4 I see xtrkcad's memory spike to over a Gb with a lot of disk noise.
Then it settles down to around 4Mb.
I'm quessing somebody is rendering the whole bitmap at the new resolution,
picking the part that maps onto the screen and throws the rest aways.
I don't know if that's Windows fault or ours.

At down to 7:1 all is well, 6:1 and 5:1 no errors but no background either (some other bug?)
Then at 4:1, I hit an uninitialized variable exception at backgnd.c:206 in
wDrawBackground on 'rotated'
FreeImage_RescaleRect is failing (it returns 0) which causes image_type to be 0,
which is not a case in the proceeding switch so rotated is not set -> boom

The fix is to bail if FreeImage_RescaleRect fails