Make sure that when an object is Modified inside Select command it is correctly logged into Undo and Redrawn properly.
cmisc.c was not calling C_FINISH if Select was current when changing commands -
but it could be a Modify underway! The C_FINISH in Select needs to call C_OK
and C_FINISH in Modify to ensure logging and completion.

We undo log Modifies even if the Modify had no change as Select has no visibility
This is also true in Cornu and Bezier.