Fix Draw many types when Sticky is set to move UndoStart earlier and apply the change during a C_CONFIRM.
This allows the Undo to be of the Draw definition, removes the Redo and
so ensures that Esc does not invalidate the Draw.

Handling of key strokes in Draw have been cleaned up - Only Enter, Space and Tab
confirm the chnage and move onto a new element, Shift-Tab rejects it.