Partial Fix #343 Error when grouping objects
The error is detected when Grouping icomplex track segments
We generate 2 different errors
- there are more than 127 track segments ("too many track segments, remove some")
- there are track segments with index > 127 ("reorder segments so tracks appear first")

The test case uses an old def'n of the turntable.
The test case is the 2nd case. Reordering the segment avoids the error
The result creates a TURNOUT definition but some paths are missing

Updated the messages and on-line help for Group