Fix #479 Problem with N gauge Tomix 1248 and 1249 turnouts.
Adjust curved turnouts so they match expectations.
Since we don't have sample to measure we are assuming the turnouts are based on
1121 and 1122 curved pieces which yields the previous 1248/1249 parts
We tweaked the angle of the diverging end points by a small amount (0.999999deg)
which allow turnout auto connect to work with the default Connect Angle value: 1.0.
In some cases you might need to increase this value. Or make adjustments to track placement
Without samples to examine this is the best we can do.

Updated 1248/1249 and added 1278/1279 with the same geometry.