FIX #527 No change detection or undo when text changed in Change Object Properties There are actually 4 bugs here.
Bug 1: Autosave after text edit even if Autosave is off
I see no evidence for this, the original file is not changed

Bug 2: Changing text doesn't mark layout as changed and Undo not enabled
DescribeUpdate() is called when changes happen in the Describe dialog
This calls UndoStart() and UndoModify() which enables Undo and marks layout as changed
But there is no call change callback for Text objects.
Instead, when the dialog is finished (UpdateDraw, inx==-1) we check if the text has changed
If so, we replace the text in the segment.
However, UndoStart/UndoModify are not called, Undo is not enabled and the layout is not marked.
The fix is copy the relevant code from DescribeUpdate

Bug 3: If text is changed and we do Undo, the original text is replaced by garbage
segPtr->u.t.string contains the text data, but it is not saved by UndoModify
We save the track object, endpts, segs and some extra arrays for polys.
We don't save u.t.text, but free it (UpdateDraw:1049).
When do Undo, we restore the segment but it's u.t.string now points to some other data
The fix is to not MyFree(u.t.string) so when to restore the seg, we point to the old string
This is a minor memory leak and should have a standard way of handling these 'bags on the side' in Undo

Bug 4: The Describe demo doesn't update the text field.
This is because UpdateDraw checks if the text object has been updated (wTextGetModified)
In playback the lower level (gtk/msw) objects are not modified so we invoke text update if 'inPlayback' as well