Fix: 541 Cannot create new prototype due to dialog box bugs
XTrkCad is complaining about the Part Number field on the Car Part dialog (which is not visible)

Fix is to only check for valid input only when the control is visible

Workaround: enter some test in the Car Part/Part Number field

Dev Note: This fix is limited to dcar.c which used a wrapper around wControlShow():
ParamControlShow().  Added a bShown field to paramData_t to track Show/HIde calls
If the control is not shown we ignore if it is invalid.

This should be extended by replacing wControlShow() with ParamControlShow() calls
elsewhere where we show/hide controls dynamically

We might want to extend this to handle invalid controls which are not active.